MVP
This is the most docker I have used/learned in a single day.

For now this will bring up PiHole using Unbound as an authorative
resolver.

This is coupled with the SearX search engine.

So you will have an authorative adblocking DNS server that heavily
caches responses and a local search engine.