Add SearX back and change web server ports.
I was thinking about the use case of running on a network that is
flagged in some sort of spam/bad behaviour database.

It is likely that Whoogle would be unable to get results from Google as
it would be marked as a bot.

This would mean you have no serch engine if you were relying on this.

For this reason I am changing back to using SearXNG for the search
engine as it aggregates from multiple sources.

I also changed the service that uses port 80 from being PiHole to
SearXNG.

It is nicer to be able to access the search engine without having to
specify a specific port.